requirements.txt
Flask
Pillow
requests
gtfs-realtime-bindings
protobuf 